Only serve from brotli folder if both a CDN host and S3 bucket are provided
Currently, if the user has a canvas_cdn.yml file, Canvas will always
append /br to the configured CDN host if the client's browser supports
brotli. This is an issue if the user has a value for bucket set in the YAML as Canvas won't build the /br directory unless it is uploading assets to S3.

This fix simply checks to see if there is an S3 bucket when performing other checks to see if the brotli S3 bucket should be used.

Test Plan:
  - Add and remove a bucket value in canvas_cdn.yml and observe the generated URLs when loading a page.